Rename Has*Listener to *Notifier and fix generator
Remove HasClickListeners.
